// frontend/src/codegen_output/models/Configuracion.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Configuracion = {
    monto_maximo_orden_def: (number | null);
    monto_maximo_pedido_def: (number | null);
    id: number;
    fecha_ultima_actualizacion: (string | null);
};